Updating file names after code refactor
