+++ backend/app/services/text_extraction.py
